Move BigInt to value::bigint module
